Replace root scripts with 2DNA versions
